@@ -1,15 +1,52 @@
-@extends('template.app')
+@extends('layouts.master')
 
-@section('content')
-    <div class="row">
-        {{-- <div class="col-md-2 d-none d-md-block bg-light"></div> --}}
-        <!-- Spacer untuk menyesuaikan isi -->
-        <div class="col-md-10">
-            <div class="container mt-3">
-                <h2>Welcome to Dashboard</h2>
-                <p>This is a responsive dashboard layout using Bootstrap.</p>
-            </div>
+@section('web-content')
+    <div class="content-wrapper">
+        <!-- Content Header (Page header) -->
+        <div class="content-header">
+            <div class="container-fluid">
+                <div class="row mb-2">
+                    <div class="col-sm-6">
+                        @if(Auth::user()->kode_role == 1)
+                            <h1 class="m-0">Selamat Datang Admin </h1>
+                        @endif
+                        @if(Auth::user()->kode_role == 2)
+                            <h1 class="m-0">Selamat Datang Prodi </h1>
+                        @endif
+                        @if(Auth::user()->kode_role == 3)
+                        <h1 class="m-0">Selamat Datang Mahasiswa </h1>
+                        @endif
+                    </div><!-- /.col -->
+                    <div class="col-sm-6">
+                        <ol class="breadcrumb float-sm-right">
+                            <li class="breadcrumb-item"><a href="#">Home</a></li>
+                            <li class="breadcrumb-item active">Dashboard</li>
+                        </ol>
+                    </div><!-- /.col -->
+                </div><!-- /.row -->
+            </div><!-- /.container-fluid -->
         </div>
+        <!-- /.content-header -->
+
+        <!-- Main content -->
+        <div class="content">
+            <div class="container-fluid">
+                
+            </div><!-- /.container-fluid -->
+        </div>
+        <!-- /.content -->
     </div>
-    </div>
+@endsection
+
+@section('ExtraCSS')
+    <link rel="stylesheet" href="{{ asset('plugins/datatables-bs4/css/dataTables.bootstrap4.min.css') }}">
+    <link rel="stylesheet" href="{{ asset('plugins/sweetalert2/sweetalert2.css') }}">
+@endsection
+
+@section('ExtraJS')
+    <script src="{{ asset('plugins/datatables/jquery.dataTables.min.js') }}"></script>
+    <script src="{{ asset('plugins/datatables-bs4/js/dataTables.bootstrap4.min.js') }}"></script>
+    <script>
+    </script>
+    <script src="{{ asset('plugins/sweetalert2/sweetalert2.js') }}"></script>
 @endsection
